+++ compose-mds/src/main/java/ch/sbb/compose_mds/composables/notificationBox/SBBNotificationBox.kt
@@ -1,4 +1,4 @@
-package ch.sbb.compose_mds.beta.notificationBox
+package ch.sbb.compose_mds.composables.notificationBox
 
 import SBBTheme
 import android.content.res.Configuration