+++ example/src/main/java/ch/sbb/compose_mds/example/pages/ColorPage.kt
@@ -13,7 +13,6 @@ import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.GridItemSpan
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -22,8 +21,8 @@ import androidx.compose.ui.graphics.toArgb
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import ch.sbb.compose_mds.beta.ExperimentalSBBComponent
-import ch.sbb.compose_mds.beta.container.SBBGroup
 import ch.sbb.compose_mds.beta.list.SBBListHeader
+import ch.sbb.compose_mds.composables.container.SBBContentBox
 import ch.sbb.compose_mds.theme.PrimitiveColors
 import ch.sbb.compose_mds.theme.SBBConst
 import ch.sbb.compose_mds.theme.SBBSpacing
@@ -88,10 +87,10 @@ private fun ColorEntry(
     name: String,
     color: Color,
 ) {
-    SBBGroup(
+    SBBContentBox(
         modifier =
-            Modifier
-                .size(100.dp),
+            Modifier.size(100.dp),
+        contentPadding = PaddingValues(0.dp),
     ) {
         Box(
             modifier =
@@ -123,13 +122,11 @@ private fun ColorEntry(
     }
 }
 
-@Preview(showBackground = true)
-@Preview(showBackground = true, uiMode = Configuration.UI_MODE_NIGHT_YES)
+@Preview(uiMode = Configuration.UI_MODE_NIGHT_NO)
+@Preview(uiMode = Configuration.UI_MODE_NIGHT_YES)
 @Composable
-private fun Preview_TypographyPage() {
-    SBBTheme {
-        Surface {
-            ColorPage()
-        }
+private fun Preview_ColorPage() {
+    SBBTheme(includeSurface = true) {
+        ColorPage()
     }
 }