+++ example/src/main/java/ch/sbb/compose_mds/example/pages/MainPage.kt
@@ -3,19 +3,20 @@ package ch.sbb.compose_mds.example.pages
 import SBBTheme
 import android.content.res.Configuration
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import androidx.navigation.compose.rememberNavController
 import ch.sbb.compose_mds.beta.ExperimentalSBBComponent
-import ch.sbb.compose_mds.beta.container.SBBGroup
 import ch.sbb.compose_mds.beta.list.SBBListHeader
 import ch.sbb.compose_mds.beta.list.SBBListItem
+import ch.sbb.compose_mds.composables.container.SBBContentBox
 import ch.sbb.compose_mds.sbbicons.SBBIcons
 import ch.sbb.compose_mds.sbbicons.Small
 import ch.sbb.compose_mds.sbbicons.small.ChevronSmallRightSmall
@@ -24,7 +25,6 @@ import ch.sbb.compose_mds.theme.defaultPadding
 @OptIn(ExperimentalSBBComponent::class)
 @Composable
 fun MainPage(navController: NavController) {
-
     Column(
         modifier = Modifier
             .defaultPadding()
@@ -34,98 +34,71 @@ fun MainPage(navController: NavController) {
             )
     ) {
         SBBListHeader(text = "Basics")
-        SBBGroup {
-            SBBListItem(
-                title = "Icon",
-                trailingIcon = SBBIcons.Small.ChevronSmallRightSmall,
-                onClick = { navController.navigate("icon") },
-            )
-            SBBListItem(
+        SBBContentBox(contentPadding = PaddingValues(0.dp)) {
+            ComponentListItem(title = "Icon", onClick = { navController.navigate("icon") })
+            ComponentListItem(
                 title = "Typography",
-                trailingIcon = SBBIcons.Small.ChevronSmallRightSmall,
                 onClick = { navController.navigate("typography") },
             )
-            SBBListItem(
+            ComponentListItem(
                 title = "Color",
-                trailingIcon = SBBIcons.Small.ChevronSmallRightSmall,
                 onClick = { navController.navigate("color") },
                 isLastElement = true,
             )
         }
         SBBListHeader(text = "Components")
-        SBBGroup {
-            SBBListItem(
-                title = "Modal View",
-                trailingIcon = SBBIcons.Small.ChevronSmallRightSmall,
-                onClick = { navController.navigate("modal-view") },
-            )
-            SBBListItem(
-                title = "Checkbox",
-                trailingIcon = SBBIcons.Small.ChevronSmallRightSmall,
-                onClick = { navController.navigate("checkbox") },
-            )
-            SBBListItem(
-                title = "Button",
-                trailingIcon = SBBIcons.Small.ChevronSmallRightSmall,
-                onClick = { navController.navigate("button") },
-            )
-            SBBListItem(
-                title = "SegmentedButton",
-                trailingIcon = SBBIcons.Small.ChevronSmallRightSmall,
-                onClick = { navController.navigate("segmented-button") },
-            )
-            SBBListItem(
+        SBBContentBox(contentPadding = PaddingValues(0.dp)) {
+            ComponentListItem(title = "Button", onClick = { navController.navigate("button") })
+            ComponentListItem(title = "Checkbox", onClick = { navController.navigate("checkbox") })
+            ComponentListItem(
+                title = "Container",
+                onClick = { navController.navigate("container") })
+            ComponentListItem(title = "Header", onClick = { navController.navigate("header") })
+            ComponentListItem(
                 title = "LoadingIndicator",
-                trailingIcon = SBBIcons.Small.ChevronSmallRightSmall,
-                onClick = { navController.navigate("loading-indicator") },
-            )
-            SBBListItem(
-                title = "Switch",
-                trailingIcon = SBBIcons.Small.ChevronSmallRightSmall,
-                onClick = { navController.navigate("switch") },
-            )
-            SBBListItem(
-                title = "Header",
-                trailingIcon = SBBIcons.Small.ChevronSmallRightSmall,
-                onClick = { navController.navigate("header") },
-            )
-            SBBListItem(
-                title = "TextField",
-                trailingIcon = SBBIcons.Small.ChevronSmallRightSmall,
-                onClick = { navController.navigate("text-field") },
-            )
-            SBBListItem(
-                title = "Status",
-                trailingIcon = SBBIcons.Small.ChevronSmallRightSmall,
-                onClick = { navController.navigate("status") },
-            )
-            SBBListItem(
-                title = "Slider",
-                trailingIcon = SBBIcons.Small.ChevronSmallRightSmall,
-                onClick = { navController.navigate("slider") },
-            )
-            SBBListItem(
+                onClick = { navController.navigate("loading-indicator") })
+            ComponentListItem(
+                title = "Modal View",
+                onClick = { navController.navigate("modal-view") })
+            ComponentListItem(
                 title = "NotificationBox",
-                trailingIcon = SBBIcons.Small.ChevronSmallRightSmall,
-                onClick = { navController.navigate("notification-box") },
-            )
-            SBBListItem(
-                title = "TabBar",
-                trailingIcon = SBBIcons.Small.ChevronSmallRightSmall,
-                onClick = { navController.navigate("tab-bar") },
+                onClick = { navController.navigate("notification-box") })
+            ComponentListItem(
+                title = "SegmentedButton",
+                onClick = { navController.navigate("segmented-button") })
+            ComponentListItem(title = "Slider", onClick = { navController.navigate("slider") })
+            ComponentListItem(title = "Status", onClick = { navController.navigate("status") })
+            ComponentListItem(title = "Switch", onClick = { navController.navigate("switch") })
+            ComponentListItem(title = "TabBar", onClick = { navController.navigate("tab-bar") })
+            ComponentListItem(
+                title = "TextField",
+                onClick = { navController.navigate("text-field") },
                 isLastElement = true,
             )
         }
     }
 }
 
-@Preview(showBackground = true)
-@Preview(showBackground = true, uiMode = Configuration.UI_MODE_NIGHT_YES)
+@OptIn(ExperimentalSBBComponent::class)
 @Composable
-fun MainPagePreview() {
-    SBBTheme {
-        Surface {
-            MainPage(navController = rememberNavController())
-        }
+private fun ComponentListItem(
+    title: String,
+    onClick: (() -> Unit)? = null,
+    isLastElement: Boolean = false,
+) {
+    SBBListItem(
+        title = title,
+        trailingIcon = SBBIcons.Small.ChevronSmallRightSmall,
+        onClick = onClick,
+        isLastElement = isLastElement
+    )
+}
+
+@Preview(uiMode = Configuration.UI_MODE_NIGHT_NO)
+@Preview(uiMode = Configuration.UI_MODE_NIGHT_YES)
+@Composable
+private fun MainPagePreview() {
+    SBBTheme(includeSurface = true) {
+        MainPage(navController = rememberNavController())
     }
 }