+++ example/src/main/java/ch/sbb/compose_mds/example/pages/ModalViewPage.kt
@@ -1,5 +1,6 @@
 package ch.sbb.compose_mds.example.pages
 
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
@@ -13,9 +14,11 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import ch.sbb.compose_mds.beta.ExperimentalSBBComponent
-import ch.sbb.compose_mds.beta.modal.SBBModalView
 import ch.sbb.compose_mds.beta.button.SBBSecondaryButton
+import ch.sbb.compose_mds.beta.list.SBBListHeader
+import ch.sbb.compose_mds.beta.modal.SBBModalView
 import ch.sbb.compose_mds.example.composeable.Placeholder
+import ch.sbb.compose_mds.theme.defaultPadding
 
 
 @OptIn(ExperimentalMaterial3Api::class, ExperimentalSBBComponent::class)
@@ -24,13 +27,16 @@ fun ModalViewPage() {
     val sheetState = rememberModalBottomSheetState()
     var showBottomSheet by remember { mutableStateOf(false) }
 
-    SBBSecondaryButton(
-        label = "Show ModalView",
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 8.dp, vertical = 8.dp),
-        onClick = { showBottomSheet = true }
-    )
+    Column(modifier = Modifier.defaultPadding()) {
+        SBBListHeader(text = "ModalView")
+        SBBSecondaryButton(
+            label = "Show",
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(horizontal = 8.dp, vertical = 8.dp),
+            onClick = { showBottomSheet = true }
+        )
+    }
 
     if (showBottomSheet) {
         SBBModalView(