+++ example/src/main/java/ch/sbb/compose_mds/example/pages/SegmentedButtonPage.kt
@@ -6,7 +6,9 @@ import android.content.res.Configuration
 import android.widget.Toast
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -17,54 +19,59 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import ch.sbb.compose_mds.beta.ExperimentalSBBComponent
-import ch.sbb.compose_mds.beta.button.SBBButtonSegment
-import ch.sbb.compose_mds.beta.container.SBBGroup
-import ch.sbb.compose_mds.beta.button.SBBSegmentedButton
+import ch.sbb.compose_mds.beta.list.SBBListHeader
+import ch.sbb.compose_mds.beta.segmentedButton.SBBButtonSegment
+import ch.sbb.compose_mds.beta.segmentedButton.SBBSegmentedButton
+import ch.sbb.compose_mds.theme.defaultPadding
 
 @OptIn(ExperimentalSBBComponent::class)
 @Composable
 fun SegmentedButtonPage() {
     val context = LocalContext.current
 
-    SBBGroup(modifier = Modifier.padding(8.dp)) {
+    Column(
+        modifier = Modifier
+            .defaultPadding()
+            .fillMaxWidth()
+            .verticalScroll(
+                state = rememberScrollState(),
+            ),
+        verticalArrangement = Arrangement.spacedBy(8.dp)
+    ) {
 
-        Column(
-            modifier = Modifier.padding(16.dp),
-            verticalArrangement = Arrangement.spacedBy(8.dp)
-        ) {
+        // DEFAULT
+        SBBListHeader(text = "Default")
 
-            // DEFAULT
+        // Default with three segments
+        var defaultSelection1: String by remember { mutableStateOf("Selection 1") }
+        SBBSegmentedButton(
+            onSelectionChanged = { selected ->
+                defaultSelection1 = selected
+                toastSelection(context, selected)
+            },
+            selection = defaultSelection1,
+            segments = threeButtonSegments()
+        )
 
-            // Default with three segments
-            var defaultSelection1: String by remember { mutableStateOf("Selection 1") }
-            SBBSegmentedButton(
-                onSelectionChanged = { selected ->
-                    defaultSelection1 = selected
-                    toastSelection(context, selected)
-                },
-                selection = defaultSelection1,
-                segments = threeButtonSegments()
-            )
+        // Default with two segments
+        var defaultSelection2: String by remember { mutableStateOf("Selection 1") }
+        SBBSegmentedButton(
+            onSelectionChanged = { selected ->
+                defaultSelection2 = selected
+                toastSelection(context, selected)
+            },
+            selection = defaultSelection2,
+            segments = twoButtonSegments()
+        )
 
-            // Default with two segments
-            var defaultSelection2: String by remember { mutableStateOf("Selection 1") }
-            SBBSegmentedButton(
-                onSelectionChanged = { selected ->
-                    defaultSelection2 = selected
-                    toastSelection(context, selected)
-                },
-                selection = defaultSelection2,
-                segments = twoButtonSegments()
-            )
-
-            // Default with two segments, disabled
-            SBBSegmentedButton(
-                onSelectionChanged = {},
-                selection = "Selection 1",
-                segments = twoButtonSegments(),
-                enabled = false,
-            )
-        }
+        // Default with two segments, disabled
+        SBBListHeader(text = "Disabled")
+        SBBSegmentedButton(
+            onSelectionChanged = {},
+            selection = "Selection 1",
+            segments = twoButtonSegments(),
+            enabled = false,
+        )
     }
 }
 
@@ -104,11 +111,11 @@ private fun toastSelection(context: Context, selection: String) {
     ).show()
 }
 
-@Preview(showBackground = true)
+@Preview(uiMode = Configuration.UI_MODE_NIGHT_NO)
 @Preview(uiMode = Configuration.UI_MODE_NIGHT_YES)
 @Composable
-fun SegmentedButtonPagePreview() {
-    SBBTheme {
+fun Preview_SegmentedButtonPage() {
+    SBBTheme(includeSurface = true) {
         SegmentedButtonPage()
     }
 }