+++ example/src/main/java/ch/sbb/compose_mds/example/pages/SliderPage.kt
@@ -2,14 +2,17 @@ package ch.sbb.compose_mds.example.pages
 
 import SBBTheme
 import android.content.res.Configuration
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.SliderState
-import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import ch.sbb.compose_mds.beta.ExperimentalSBBComponent
-import ch.sbb.compose_mds.beta.container.SBBGroup
+import ch.sbb.compose_mds.beta.list.SBBListHeader
 import ch.sbb.compose_mds.beta.slider.SBBSlider
 import ch.sbb.compose_mds.theme.defaultPadding
 import ch.sbb.compose_mds.theme.elementPadding
@@ -17,22 +20,27 @@ import ch.sbb.compose_mds.theme.elementPadding
 @OptIn(ExperimentalMaterial3Api::class, ExperimentalSBBComponent::class)
 @Composable
 fun SliderPage() {
-    SBBGroup(
-        modifier = Modifier.defaultPadding()
+    Column(
+        modifier = Modifier
+            .defaultPadding()
+            .fillMaxWidth()
+            .verticalScroll(
+                state = rememberScrollState(),
+            ),
     ) {
         val sliderState = SliderState(value = .25f)
+        SBBListHeader(text = "Default")
         SBBSlider(modifier = Modifier.elementPadding(), state = sliderState)
+        SBBListHeader(text = "Disabled")
         SBBSlider(modifier = Modifier.elementPadding(), state = sliderState, enabled = false)
     }
 }
 
-@Preview(showBackground = true)
-@Preview(showBackground = true, uiMode = Configuration.UI_MODE_NIGHT_YES)
+@Preview(uiMode = Configuration.UI_MODE_NIGHT_NO)
+@Preview(uiMode = Configuration.UI_MODE_NIGHT_YES)
 @Composable
-private fun Preview_SliderPage() {
-    SBBTheme {
-        Surface {
-            SliderPage()
-        }
+fun Preview_SliderPage() {
+    SBBTheme(includeSurface = true) {
+        SliderPage()
     }
 }