+++ example/src/main/java/ch/sbb/compose_mds/example/pages/TabBarPage.kt
@@ -7,7 +7,6 @@ import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.navigationBars
 import androidx.compose.foundation.layout.windowInsetsPadding
-import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
@@ -59,13 +58,11 @@ fun TabBarPage() {
     }
 }
 
-@Preview(showBackground = true)
-@Preview(showBackground = true, uiMode = Configuration.UI_MODE_NIGHT_YES)
+@Preview(uiMode = Configuration.UI_MODE_NIGHT_NO)
+@Preview(uiMode = Configuration.UI_MODE_NIGHT_YES)
 @Composable
-private fun Preview_TabBarPage() {
-    SBBTheme {
-        Surface {
-            TabBarPage()
-        }
+fun Preview_TabBarPage() {
+    SBBTheme(includeSurface = true) {
+        TabBarPage()
     }
 }