+++ example/src/main/java/ch/sbb/compose_mds/example/pages/TypographyPage.kt
@@ -3,23 +3,23 @@ package ch.sbb.compose_mds.example.pages
 import SBBTheme
 import android.content.res.Configuration
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.tooling.preview.Preview
 import ch.sbb.compose_mds.beta.ExperimentalSBBComponent
-import ch.sbb.compose_mds.beta.container.SBBGroup
 import ch.sbb.compose_mds.beta.list.SBBListHeader
+import ch.sbb.compose_mds.composables.container.SBBContentBox
 import ch.sbb.compose_mds.theme.SBBConst
 import ch.sbb.compose_mds.theme.SBBSpacing
-import ch.sbb.compose_mds.theme.SBBTypography
+import ch.sbb.compose_mds.theme.defaultPadding
 
 private const val DEFAULT_TEXT = "The quick brown fox jumps over the lazy dog"
 
@@ -28,8 +28,8 @@ fun TypographyPage() {
     Column(
         modifier =
             Modifier
-                .verticalScroll(rememberScrollState())
-                .padding(vertical = SBBSpacing.XSmall),
+                .defaultPadding()
+                .verticalScroll(rememberScrollState()),
     ) {
         MaterialTypography()
         SBBTypographySection()
@@ -77,8 +77,6 @@ private fun SBBTypographySection() {
     TextStylePreview(name = "XSmallBold", SBBTheme.sbbTypography.XSmallBold)
     TextStylePreview(name = "XXSmallLight", SBBTheme.sbbTypography.XXSmallLight)
     TextStylePreview(name = "XXSmallBold", SBBTheme.sbbTypography.XXSmallBold)
-
-
 }
 
 @OptIn(ExperimentalSBBComponent::class)
@@ -88,7 +86,7 @@ private fun TextStylePreview(
     style: TextStyle,
 ) {
     SBBListHeader(text = name)
-    SBBGroup(modifier = Modifier.padding(horizontal = SBBConst.DEFAULT_HORIZONTAL_PADDING)) {
+    SBBContentBox(contentPadding = PaddingValues(horizontal = SBBConst.DEFAULT_HORIZONTAL_PADDING)) {
         Text(
             text = DEFAULT_TEXT,
             style = style,
@@ -108,13 +106,11 @@ private fun TextStyleInfos(style: TextStyle) {
     )
 }
 
-@Preview(showBackground = true)
-@Preview(showBackground = true, uiMode = Configuration.UI_MODE_NIGHT_YES)
+@Preview(uiMode = Configuration.UI_MODE_NIGHT_NO)
+@Preview(uiMode = Configuration.UI_MODE_NIGHT_YES)
 @Composable
-private fun Preview_TypographyPage() {
-    SBBTheme {
-        Surface {
-            TypographyPage()
-        }
+fun Preview_TypographyPage() {
+    SBBTheme(includeSurface = true) {
+        TypographyPage()
     }
 }