+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+# Unversioned
+
+### Features
+* **SBBContentBox:** added component which replaces SBBGroup
+
 # 0.1.8
 
 ### Bugfix